Fix jumping IMU data issue

+++ zeropilot4.0/include/attitude_manager/attitude_manager.hpp
@@ -71,7 +71,7 @@ class AttitudeManager {
 
         static constexpr uint16_t MAX_TIMESTAMP = 65535;
         static constexpr float TIMESTAMP_RESOLUTION = 0.000001f; // Default IMU timestamp resolution 1us
-        uint16_t lastTimestamp;
+        uint32_t lastTimestamp;
         bool haveLastImuTimestamp;
 
         bool getControlInputs(RCMotorControlMessage_t *pControlMsg);
